backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==5.0.3
django-cors-headers==4.3.1
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.3
fonttools==4.50.0
idna==3.6
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
parso==0.8.3
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-whois==0.9.3
pytz==2024.1
pywin32==306
pyzmq==25.1.2
requests==2.31.0
requests-file==2.0.0
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
threadpoolctl==3.4.0
tldextract==5.1.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
